Fix Lambda handler, and move to runner package
Tool: gitpod/catfood.gitpod.cloud
